pyproject.toml
[project]
name = "rag-telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "faiss-cpu>=1.12.0",
    "numpy>=2.2.6",
    "pypdf2>=3.0.1",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
]
